+++ packages/core/__tests__/onDocumentLoad.test.tsx
@@ -3,17 +3,25 @@ import { render } from '@testing-library/react';
 
 import { mockIsIntersecting } from '../../../test-utils/mockIntersectionObserver';
 import { Viewer } from '../src/Viewer';
+import { DocumentLoadEvent } from '../src';
 
 const TestOnDocumentLoad: React.FC<{
     fileUrl: Uint8Array;
 }> = ({ fileUrl }) => {
     const [numPages, setNumPages] = React.useState(0);
+    const [fileLength, setFileLength] = React.useState(0);
+
+    const handleDocumentLoad = (e: DocumentLoadEvent) => {
+        setNumPages(e.doc.numPages);
+        setFileLength(e.file.data.length);
+    };
 
     return (
         <>
+            <div data-testid="file-length">{fileLength}</div>
             <div data-testid="num-pages">{numPages}</div>
             <div style={{ height: '720px' }}>
-                <Viewer fileUrl={fileUrl} onDocumentLoad={(e) => setNumPages(e.doc.numPages)} />
+                <Viewer fileUrl={fileUrl} onDocumentLoad={handleDocumentLoad} />
             </div>
         </>
     );
@@ -25,4 +33,7 @@ test('onDocumentLoad() callback', async () => {
 
     const numPagesLabel = await findByTestId('num-pages');
     expect(numPagesLabel.textContent).toEqual('2');
+
+    const fileLengthLabel = await findByTestId('file-length');
+    expect(fileLengthLabel.innerHTML).toEqual('3028');
 });