+++ packages/core/src/types/Plugin.ts
@@ -9,7 +9,8 @@
 import * as React from 'react';
 
 import { LayerRenderStatus } from '../structs/LayerRenderStatus';
-import type { PdfJs } from '../types/PdfJs';
+import type { OpenFile } from './OpenFile';
+import type { PdfJs } from './PdfJs';
 import type { PluginFunctions } from './PluginFunctions';
 import type { RenderViewer } from './RenderViewer';
 import type { Slot } from './Slot';
@@ -17,6 +18,7 @@ import type { ViewerState } from './ViewerState';
 
 export interface PluginOnDocumentLoad {
     doc: PdfJs.PdfDocument;
+    file: OpenFile;
 }
 
 export interface PluginOnTextLayerRender {